backend/requirements.txt
### Core RAG stack
weaviate-client
docling
llama-index

### API and web
fastapi[standard]
uvicorn

### Embeddings / rerankers
rank-bm25
cohere

### Semantic cache (Redis Stack with RediSearch required for vector search)
redis
numpy

### Configuration / validation
pydantic
pydantic-settings

### Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
pytest-timeout>=2.1.0
httpx>=0.24.0
faker>=18.0.0